Only update the vector clock if it's not nil
This fixes an issue where it appeared different
environments would treat a type ambiguity differently.
In one environment (travis-ci) the nil
vector clock would be treated as a byte-array.
In my local testing it was treated as a VClock
object. This lead to an exception in the travi-ci
environment. In this case, we skip the ambiguity
by always treating the vector clock as a byte-array
on return, and only updating it if it's not nil.